Refactor : Added barrel export for components
