+++ src/barrelexport/Pageutil.js
@@ -1,0 +1,7 @@
+export { Card } from "../components/Card/Card";
+export { Login } from "../pages/Authentication/Login/Login";
+export { Signup } from "../pages/Authentication/Signup/Signup";
+export { Cart } from "../pages/Cart/Cart";
+export { Home } from "../pages/Home/Home";
+export { Products } from "../pages/Products/Products";
+export { Wishlist } from "../pages/Wishlist/Wishlist";
