+++ src/pages/Authentication/Signup/Signup.jsx
@@ -1,8 +1,7 @@
 import React from "react";
-import { AiOutlineMail, AiOutlineLock } from "react-icons/ai";
 import { FaRegEyeSlash } from "react-icons/fa";
 import { Link } from "react-router-dom";
-import { Navbar } from "../../../components/Navbar/Navbar";
+import { Navbar } from "../../../barrelexport/Componentutil";
 import "./signup.css";
 
 export const Signup = () => {
