--- src/pages/Pageutil.js
@@ -1,7 +1,0 @@
-export { Card } from "../components/Card/Card";
-export { Login } from "./Authentication/Login/Login";
-export { Signup } from "./Authentication/Signup/Signup";
-export { Cart } from "./Cart/Cart";
-export { Home } from "./Home/Home";
-export { Products } from "./Products/Products";
-export { Wishlist } from "./Wishlist/Wishlist";
